requirements.txt
# Persistent Character Agent
# Python Dependencies

# API Client - For Google Gemini API
requests>=2.28.0

# Semantic Search & Embeddings
llama-cpp-python>=0.3.0
faiss-cpu>=1.7.4
numpy>=1.21.0

# Note: If you have a GPU, use faiss-gpu instead of faiss-cpu:
# faiss-gpu>=1.7.4
